build(Gradle): Start migrating scanner plugins to separate Gradle projects
For now, only migrate the simple CLI-based scanners. The remaining
scanners require more refactoring before they can be migrated.

This change requires to use a dedicated test license file bundled as a
resource to be independent of the file location of the test that derives
from `AbstractPathScannerWrapperFunTest`.
