+++ plugins/scanners/askalono/src/main/resources/META-INF/services/org.ossreviewtoolkit.scanner.ScannerWrapperFactory
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.scanners.askalono.Askalono$Factory
