+++ plugins/scanners/boyterlc/src/main/resources/META-INF/services/org.ossreviewtoolkit.scanner.ScannerWrapperFactory
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.scanners.boyterlc.BoyterLc$Factory
