+++ scanner/src/main/resources/META-INF/services/org.ossreviewtoolkit.scanner.ScannerWrapperFactory
@@ -1,6 +1,3 @@
-org.ossreviewtoolkit.scanner.scanners.Askalono$Factory
-org.ossreviewtoolkit.scanner.scanners.BoyterLc$Factory
-org.ossreviewtoolkit.scanner.scanners.Licensee$Factory
 org.ossreviewtoolkit.scanner.scanners.fossid.FossId$Factory
 org.ossreviewtoolkit.scanner.scanners.scancode.ScanCode$Factory
 org.ossreviewtoolkit.scanner.scanners.scanoss.ScanOss$Factory
