+++ scanner/src/testFixtures/kotlin/AbstractPathScannerWrapperFunTest.kt
@@ -53,11 +53,10 @@ abstract class AbstractPathScannerWrapperFunTest(testTags: Set<Tag> = emptySet()
     override suspend fun beforeSpec(spec: Spec) {
         inputDir = tempdir()
 
-        // Copy our own root license under different names to a temporary directory, so we have something to operate on.
-        val ortLicense = File("../LICENSE")
+        // Create variants of the Apache-2.0 license text in a temporary directory, so we have something to operate on.
+        val licenseText = javaClass.getResource("/LICENSE").readText()
         commonlyDetectedFiles.forEach {
-            val text = ortLicense.readText()
-            inputDir.resolve(it).writeText(text.replace("license", it))
+            inputDir.resolve(it).writeText(licenseText.replace("license", it))
         }
     }
